Placed chargen functions in own file
