+++ scr/generic.py
@@ -1,0 +1,8 @@
+import os
+
+
+def clear():
+    if os.name == 'nt':
+        os.system('cls')
+    else:
+        os.system('clear')
